added a mouse enter function
The pictures were being loaded every time the ouse moved, Edge was taking it seriously and not checking the cache or something.
Added a mouse enter function to load the pictures once, and mouse move know works normally but not loading pictures by itself. It isn't bothered by the loading of attributes at all. This saves the requests for no reason at all when I'm consuming stale data anyway.